index tags from redis
The tags in redis are messy. This drops two types of problem tag:

* tags with string encoded unicode characters (`\u00e9`)
* tags with `"` in them

In both cases ES chokes when trying to index a document with those tags.
Because we're doing it in bulk, it misses 999 other tagged documents.

TODO: properly encode unicode and quotes so those tags make it through.